+++ packages/ls-components/components/confirm/types.ts
@@ -70,6 +70,14 @@ export const lsConfirmProps = buildProps({
   requestParams: {
     type: Object,
     default: () => ({})
+  },
+  icon: {
+    type: Object,
+    default: () => ({})
+  },
+  closeIcon: {
+    type: Object,
+    default: null
   }
 });