+++ packages/ls-components/components/tree/types.ts
@@ -31,11 +31,11 @@ export const lsTreeProps = buildProps({
     type: Boolean,
     default: false
   },
-  // 展示方式 默认：垂直展示
-  direction: {
-    type: String as PropType<'h' | 'v'>,
-    default: 'v'
-  },
+  // // 展示方式 默认：垂直展示
+  // direction: {
+  //   type: String as PropType<'h' | 'v'>,
+  //   default: 'v'
+  // },
   showCheckbox: {
     type: Boolean,
     default: true
@@ -53,12 +53,16 @@ export const lsTreeProps = buildProps({
     default: false
   },
   treeData: {
-    type: Array,
+    type: Array<any>,
     default: () => []
   },
   defaultCheckedKeys: {
     type: Array<any>,
     default: () => []
+  },
+  attrs: {
+    type: Object,
+    default: () => ({})
   }
 });