+++ packages/ls-components/src/views/list/Index.vue
@@ -28,8 +28,7 @@ const formItems = [
 const tableColumn = [
   {
     label: '名称',
-    prop: 'name',
-    minWidth: 120
+    prop: 'name'
   },
   {
     label: '类型',
@@ -39,19 +38,17 @@ const tableColumn = [
       A: { type: 'success', label: '类型A' },
       B: { type: '', label: '类型B' }
     },
-    minWidth: 120,
     statusStyle: 'follow'
   },
   {
     label: '自定义',
     prop: 'table-slot',
-    type: 'slot',
-    minWidth: 120
+    type: 'slot'
   },
   {
     label: '创建时间',
     prop: 'createTime',
-    minWidth: 120
+    type: 'date'
   }
 ];
 
@@ -66,7 +63,7 @@ function generateTableData(pageNum: number, pageSize: number) {
       id: index,
       name: `测试数据${index}`,
       type: index % 2 === 0 ? 'A' : 'B',
-      createTime: '2024-01-01',
+      createTime: 1740000000000 + index * 1000 * 60 * 60 * 24,
       status: index % 2
     });
   }
@@ -82,10 +79,16 @@ function listApi() {
   });
 }
 
+function dealParams(params: any) {
+  return {
+    ...params
+  };
+}
+
 function dealData(res: any) {
   const list = (res || []).map((item: any) => {
     item.popconfirmTxt = `确定删除该记录：${item.name}？`;
-    item.tableDetailText = `${item.name}详情`;
+    // item.tableDetailText = `${item.name}详情`;
     return item;
   });
   return {
@@ -99,30 +102,120 @@ const spacer = h(ElDivider, { direction: 'vertical' });
 function getTableDelText(row: any) {
   return `删除-${row.name}`;
 }
+
+function queryFn(val: any) {
+  console.log('queryFn', val);
+}
+
+const selection = ref<any[]>([
+  {
+    id: 2,
+    name: '测试数据2',
+    type: 'A',
+    createTime: '2024-01-01'
+  }
+]);
+
+watch(
+  () => selection.value,
+  newVal => {
+    console.log('watch---选中数据', newVal);
+  },
+  {
+    immediate: true,
+    deep: true
+  }
+);
+
+function updateSelection(val: any) {
+  console.log('updateSelection---选中数据', val);
+  selection.value = val;
+}
+
+function delApi(id: any) {
+  return new Promise(resolve => {
+    setTimeout(() => {
+      resolve({
+        code: 200,
+        id
+      });
+    }, 1000);
+  });
+}
+
+// function tableDelFn(row: any, setLoading: (loading: boolean) => void) {
+//   ElMessage.success(row?.name);
+//   setLoading(true);
+
+//   setTimeout(() => {
+//     setLoading(false);
+//   }, 1000);
+// }
+
+function delSuccess(row: any, res: boolean) {
+  console.log('delSuccess', row, res);
+}
+
+function resetForm(val: any) {
+  console.log(val);
+}
 </script>
 
 <template>
   <LSList
     ref="ListRef"
     :list-api="listApi"
+    :del-api="delApi"
+    :del-message="(row: any) => `hhhhhh----${row.name}`"
     :form-data="formData"
     :form-items="formItems"
     :table-column="tableColumn"
-    :show-table-switch="false"
+    :show-table-switch="true"
     :show-table-edit="true"
     :disabled-table-del="(row: any) => row.name === '测试1'"
     :disabled-table-switch="(row: any) => row.name === '测试1'"
     :deal-data="dealData"
+    :deal-params="dealParams"
     :table-operate-column="{ minWidth: 200 }"
-    :table-attrs="{ showSelect: true }"
+    :table-attrs="{
+      showSelect: true,
+      selection,
+      rowKey: 'id',
+      selectColumnOptions: {
+        // reserveSelection: false
+        'reserve-selection': true
+      },
+      // 'show-overflow-tooltip': false
+      'show-overflow-tooltip': {
+        'popper-class': 'red'
+      }
+    }"
+    :table-listeners="{
+      'selection-change': (val: any) => {
+        console.log('selection-change', val);
+      },
+      'update:selection': updateSelection
+    }"
     :show-table-operate="true"
     :disabled-add-btn="true"
-    :list-hook-config="{ currentPageProp: 1, pageSizeProp: 10, pageSizesProp: [10, 20, 30, 40, 50, 100] }"
+    :list-hook-config="{
+      currentPageProp: 1,
+      pageSizeProp: 10,
+      pageSizesProp: [10, 20, 30, 40, 50, 100]
+    }"
     :table-detail-text="`详情`"
     :table-edit-text="(row: any) => `编辑-${row.name}`"
     :table-del-text="getTableDelText"
     :popconfirm-txt="(row: any) => `删除该数据：${row.name}？`"
+    :query-fn="queryFn"
+    :table-edit-type="(row: any) => (row.id === 1 ? 'danger' : '')"
     add-btn-text="新增数据"
+    table-switch-pop-txt="新增数据"
+    :table-switch-pop-attrs="{
+      title: 'aaa'
+    }"
+    @reset-form="resetForm"
+    @del-success="delSuccess"
   >
     <template #buttons-prepend-form-slot>
       <div>按钮区域前置</div>
@@ -158,11 +251,11 @@ function getTableDelText(row: any) {
       </el-space>
     </template>
 
-    <template #table-append>
+    <!-- <template #table-append>
       <el-table-column label="新增" width="100">
         <el-button link type="primary"> 新加 </el-button>
       </el-table-column>
-    </template>
+    </template> -->
   </LSList>
 </template>