+++ packages/ls-components/CHANGELOG.md
@@ -1,3 +1,299 @@
+# [2.2.0](https://github.com/DTFN/ls-components-plus/compare/v2.1.3...v2.2.0) (2025-10-16)
+
+
+### Bug Fixes
+
+* **ls-component:** 🧩 update ([596c23a](https://github.com/DTFN/ls-components-plus/commit/596c23a7c5a8ce81e7eeb7ed0574c0d17dfdf919))
+* **ls-component:** 🧩 update ([c197898](https://github.com/DTFN/ls-components-plus/commit/c197898345fa121efdf4408db98b58e75555a337))
+* **ls-component:** 🧩 update ([a0c36f2](https://github.com/DTFN/ls-components-plus/commit/a0c36f2567c4b880964a1559f425207d4f621df3))
+* **ls-component:** 🧩 update ([86820c9](https://github.com/DTFN/ls-components-plus/commit/86820c9c94ec85ded18a51afc62e41c8cf0c3d6f))
+* **ls-component:** 🧩 update ([7ee4378](https://github.com/DTFN/ls-components-plus/commit/7ee43789fbe27cd6c27944669d3bda867769e05d))
+
+
+### Features
+
+* **ls-component:** 🚀 update ([f6b6dfb](https://github.com/DTFN/ls-components-plus/commit/f6b6dfb0f74a6834978d41c6728b7d256e77ef0d))
+* **ls-component:** 🚀 update ([a939b04](https://github.com/DTFN/ls-components-plus/commit/a939b04b7bb0f18792a04f72b0746b42aa23daec))
+* **ls-component:** 🚀 预览加水印 ([05d1230](https://github.com/DTFN/ls-components-plus/commit/05d123053a300652abe94d8354ae67d391e27ffc))
+* **ls-component:** 🚀 预览增加水印功能 ([1a7387f](https://github.com/DTFN/ls-components-plus/commit/1a7387f68764df85112e653bc3131378fdd98907))
+
+# [2.2.0](https://github.com/DTFN/ls-components-plus/compare/v2.1.3...v2.2.0) (2025-10-16)
+
+
+### Bug Fixes
+
+* **ls-component:** 🧩 update ([596c23a](https://github.com/DTFN/ls-components-plus/commit/596c23a7c5a8ce81e7eeb7ed0574c0d17dfdf919))
+* **ls-component:** 🧩 update ([c197898](https://github.com/DTFN/ls-components-plus/commit/c197898345fa121efdf4408db98b58e75555a337))
+* **ls-component:** 🧩 update ([a0c36f2](https://github.com/DTFN/ls-components-plus/commit/a0c36f2567c4b880964a1559f425207d4f621df3))
+* **ls-component:** 🧩 update ([86820c9](https://github.com/DTFN/ls-components-plus/commit/86820c9c94ec85ded18a51afc62e41c8cf0c3d6f))
+* **ls-component:** 🧩 update ([7ee4378](https://github.com/DTFN/ls-components-plus/commit/7ee43789fbe27cd6c27944669d3bda867769e05d))
+
+
+### Features
+
+* **ls-component:** 🚀 update ([a939b04](https://github.com/DTFN/ls-components-plus/commit/a939b04b7bb0f18792a04f72b0746b42aa23daec))
+* **ls-component:** 🚀 预览增加水印功能 ([1a7387f](https://github.com/DTFN/ls-components-plus/commit/1a7387f68764df85112e653bc3131378fdd98907))
+
+## [2.1.3](https://github.com/DTFN/ls-components-plus/compare/v2.1.2...v2.1.3) (2025-08-20)
+
+
+### Bug Fixes
+
+* **ls-component:** 🧩 list组件update ([02d9171](https://github.com/DTFN/ls-components-plus/commit/02d917169619982dd67277a9fe0ced5e3d672ae2))
+* **ls-doc:** 🧩 文档更新 ([9917583](https://github.com/DTFN/ls-components-plus/commit/991758324b2cd79342112ccf303395eb83c41c44))
+
+## [2.1.2](https://github.com/DTFN/ls-components-plus/compare/v2.1.1...v2.1.2) (2025-08-13)
+
+
+### Bug Fixes
+
+* **ls-component:** 🧩 element-plus升级 ([b8d8b2d](https://github.com/DTFN/ls-components-plus/commit/b8d8b2d8bf4eb1f0511145d326f1416c77ff7a31))
+
+## [2.1.1](https://github.com/DTFN/ls-components-plus/compare/v2.1.0...v2.1.1) (2025-08-13)
+
+
+### Bug Fixes
+
+* 🧩 update ([724f64b](https://github.com/DTFN/ls-components-plus/commit/724f64b2f023135413833464da3d1e9bd7a27519))
+* **ls-component:** 🧩 decriptions label-width更新 ([de28e98](https://github.com/DTFN/ls-components-plus/commit/de28e985d6692cc6b1287c9713f107585b9d4a94))
+* **ls-component:** 🧩 description 更新 ([9c6051b](https://github.com/DTFN/ls-components-plus/commit/9c6051b5e2c1905047919a9823c0571008a2a706))
+* **ls-component:** 🧩 el-menu新属性测试 ([703344a](https://github.com/DTFN/ls-components-plus/commit/703344a07048792c0209707532fcb36beee04ab2))
+* **ls-component:** 🧩 update ([3b1a5d7](https://github.com/DTFN/ls-components-plus/commit/3b1a5d71b0433338240a0b0e8e0479bf9a675b02))
+* **ls-component:** 🧩 update ([829183e](https://github.com/DTFN/ls-components-plus/commit/829183e19781126c1192d3a38c942337737380c2))
+* **ls-component:** 🧩 公共样式污染问题处理 ([00d04ce](https://github.com/DTFN/ls-components-plus/commit/00d04ce7f6421c659921a34a8d8382495115852d))
+
+# [2.1.0](https://github.com/DTFN/ls-components-plus/compare/v2.0.3...v2.1.0) (2025-08-11)
+
+
+### Bug Fixes
+
+* **ls-component:** 🧩 chart增加titleComponent ([3c35c4d](https://github.com/DTFN/ls-components-plus/commit/3c35c4d6f7599950b4e264879f7cc71ae658fdab))
+* **ls-component:** 🧩 description数值为0时，是否展示0 ([90cd799](https://github.com/DTFN/ls-components-plus/commit/90cd799a129457461515f6dbebd534a01119bb9f))
+* **ls-component:** 🧩 update ([e1daa07](https://github.com/DTFN/ls-components-plus/commit/e1daa077a0583dfeed5eae5a0f5345b78b1f9bac))
+* **ls-component:** 🧩 update ([99b7354](https://github.com/DTFN/ls-components-plus/commit/99b735416cc704fe3a5a62e978bc1a700a4b34d9))
+* **ls-component:** 🧩 update ([2aa8cb6](https://github.com/DTFN/ls-components-plus/commit/2aa8cb63c1f9e19c03a2f2d5f4e90cf23fc09e5f))
+* **ls-component:** 🧩 update ([b4aaaea](https://github.com/DTFN/ls-components-plus/commit/b4aaaea6007a20d2511d4e833db64dfd5bbae80f))
+* **ls-component:** 🧩 update ([4330d99](https://github.com/DTFN/ls-components-plus/commit/4330d992daf0981a8d3c776c677dee265a6577b2))
+* **ls-component:** 🧩 update ([e91fa92](https://github.com/DTFN/ls-components-plus/commit/e91fa92c795da42c179856f3842286b67ee8c81b))
+* **ls-component:** 🧩 update ([d33f826](https://github.com/DTFN/ls-components-plus/commit/d33f826fa378ba0696bb634a181ff6b3825cb6e0))
+* **ls-component:** 🧩 展示注销dts生成 ([169a230](https://github.com/DTFN/ls-components-plus/commit/169a23086464324785ceeb46ebb0378d3e84dd2f))
+* **ls-component:** 🧩 文档更新 ([416ab15](https://github.com/DTFN/ls-components-plus/commit/416ab15bbc5a7aacd6c34ec4d9bf30f04dbaeb42))
+* **ls-component:** 🧩 组件升级 ([8c885a6](https://github.com/DTFN/ls-components-plus/commit/8c885a63269e97e2bda77d8651db7a0fd8f17b3c))
+
+
+### Features
+
+* **ls-component:** 🚀 element-plus升级 ([b1ffe06](https://github.com/DTFN/ls-components-plus/commit/b1ffe06e246bbeb5d882bbe80018ab4ace09ff13))
+* **ls-component:** 🚀 新api测试 ([592e47b](https://github.com/DTFN/ls-components-plus/commit/592e47bbca19cac359e36fe4a34424acbac4d2c5))
+
+## [2.0.3](https://github.com/DTFN/ls-components-plus/compare/v2.0.2...v2.0.3) (2025-07-21)
+
+
+### Bug Fixes
+
+* **ls-component:** 🧩 image预览支持点击空白区域关闭 ([74c96ed](https://github.com/DTFN/ls-components-plus/commit/74c96ed39912463ee3342d5f817ab052e365a1df))
+
+## [2.0.2](https://github.com/DTFN/ls-components-plus/compare/v2.0.1...v2.0.2) (2025-07-17)
+
+
+### Bug Fixes
+
+* **ls-component:** 🧩 upload组件，图片模式取消上传时样式优化 ([b3a296b](https://github.com/DTFN/ls-components-plus/commit/b3a296bc460b6ae9441c9ef6d64cb66a334fded9))
+
+## [2.0.1](https://github.com/DTFN/ls-components-plus/compare/v2.0.0...v2.0.1) (2025-07-16)
+
+
+### Bug Fixes
+
+* **ls-component:** 🧩 git push ([40a6f67](https://github.com/DTFN/ls-components-plus/commit/40a6f67d18f7d88e315fa4d05e07a16ff4b7cd05))
+* **ls-component:** 🧩 pdf增加初始是否分页开关 ([ba3794a](https://github.com/DTFN/ls-components-plus/commit/ba3794a8c5ee7e85ce22fe046e7e8f3f7a499244))
+
+# [2.0.0](https://github.com/DTFN/ls-components-plus/compare/v1.8.11...v2.0.0) (2025-07-15)
+
+
+### Bug Fixes
+
+* **ls-component:** 🧩 xlsx、pdf 点击空白区域支持关闭 ([e3c99a3](https://github.com/DTFN/ls-components-plus/commit/e3c99a387fe6db90af24eae70a07ba63489caf00))
+
+
+### BREAKING CHANGES
+
+* **ls-component:** xlsx、pdf 点击空白区域支持关闭
+
+## [1.8.11](https://github.com/DTFN/ls-components-plus/compare/v1.8.10...v1.8.11) (2025-07-14)
+
+
+### Bug Fixes
+
+* **ls-component:** 🧩 update ([9838713](https://github.com/DTFN/ls-components-plus/commit/9838713815afea48747fc698b77574e93cf6d592))
+* **ls-component:** 🧩 update ([ab64e44](https://github.com/DTFN/ls-components-plus/commit/ab64e448d6e9e7ac6760d388ec6bb903613f9323))
+* **ls-component:** 🧩 xlsx打开渲染显示，资源异常预览关闭及提示修复 ([e971a9a](https://github.com/DTFN/ls-components-plus/commit/e971a9a9625cb29a18f0b7b1b9bf0c481abb30a2))
+
+## [1.8.10](https://github.com/DTFN/ls-components-plus/compare/v1.8.9...v1.8.10) (2025-06-20)
+
+
+### Bug Fixes
+
+* **ls-component:** 🧩 tree组件更新 ([d4f8a3a](https://github.com/DTFN/ls-components-plus/commit/d4f8a3adc13ec94d026dd57138dbc0ea2b9c46c6))
+* **ls-component:** 🧩 update ([83169ac](https://github.com/DTFN/ls-components-plus/commit/83169ac621f0912012229d01bf8921df16636293))
+
+## [1.8.9](https://github.com/DTFN/ls-components-plus/compare/v1.8.8...v1.8.9) (2025-06-19)
+
+
+### Bug Fixes
+
+* **ls-component:** 🧩 confirm更新icon ([45a695a](https://github.com/DTFN/ls-components-plus/commit/45a695a412677c5c6f9aae92c3495376aec353d9))
+
+## [1.8.8](https://github.com/DTFN/ls-components-plus/compare/v1.8.7...v1.8.8) (2025-06-03)
+
+
+### Bug Fixes
+
+* **ls-component:** 🧩 update ([f27d716](https://github.com/DTFN/ls-components-plus/commit/f27d716c8f27e56478698e25a41bee1274a492ae))
+* **ls-component:** 🧩 文档更新 ([40503ab](https://github.com/DTFN/ls-components-plus/commit/40503ab9ee7067d4f5b5e08b50365123b9c2c00a))
+
+## [1.8.7](https://github.com/DTFN/ls-components-plus/compare/v1.8.6...v1.8.7) (2025-05-29)
+
+
+### Bug Fixes
+
+* **ls-component:** 🧩 优化dialog按钮 ([8d1578f](https://github.com/DTFN/ls-components-plus/commit/8d1578f59834bc560d255a7680d63c46274d6fe4))
+
+## [1.8.6](https://github.com/DTFN/ls-components-plus/compare/v1.8.5...v1.8.6) (2025-05-16)
+
+
+### Bug Fixes
+
+* **ls-component:** 🧩 chart增加组件 ([72e698e](https://github.com/DTFN/ls-components-plus/commit/72e698e94f7c511a6d707b2f92493eb5da4477c0))
+
+## [1.8.5](https://github.com/DTFN/ls-components-plus/compare/v1.8.4...v1.8.5) (2025-05-14)
+
+
+### Bug Fixes
+
+* **ls-components:** table组件多选优化 ([1763e1d](https://github.com/DTFN/ls-components-plus/commit/1763e1dd762f6d378be3faca1b81a65300f9593c))
+
+## [1.8.4](https://github.com/DTFN/ls-components-plus/compare/v1.8.3...v1.8.4) (2025-05-06)
+
+
+### Bug Fixes
+
+* **ls-component:** 🧩 layout完善无command数据时，展示内容 ([71e640c](https://github.com/DTFN/ls-components-plus/commit/71e640ca14e8f43ca6488a2511dee19f7a7b9754))
+* **ls-component:** 🧩 优化pdf分页按钮样式 ([08feb56](https://github.com/DTFN/ls-components-plus/commit/08feb5627103b253188fa640e9dd411056fab3f7))
+* **ls-component:** 🧩 文档更新 ([c8a741f](https://github.com/DTFN/ls-components-plus/commit/c8a741ffeef1ecc393ecfa103b233edbd0075d1c))
+* **ls-doc:** 🧩 文档更新 ([d0d3165](https://github.com/DTFN/ls-components-plus/commit/d0d31651c77df05a6a37f253f5fcc3146a3ce782))
+
+## [1.8.3](https://github.com/DTFN/ls-components-plus/compare/v1.8.2...v1.8.3) (2025-04-28)
+
+
+### Bug Fixes
+
+* **ls-component:** 🧩 layout增加title插槽 ([c1964ad](https://github.com/DTFN/ls-components-plus/commit/c1964aded0a46452f954d8e5009589535061c3ab))
+* **ls-component:** list组件queryFn函数优化 ([3f27205](https://github.com/DTFN/ls-components-plus/commit/3f27205f4caaf9cccec7f66d14dffb532ca9f003))
+
+## [1.8.2](https://github.com/DTFN/ls-components-plus/compare/v1.8.1...v1.8.2) (2025-04-24)
+
+
+### Bug Fixes
+
+* **ls-component:** 🧩 layout样式优化 ([69937d8](https://github.com/DTFN/ls-components-plus/commit/69937d81b5c99bb7e3202acd5ae4d58e8def6c3b))
+
+## [1.8.1](https://github.com/DTFN/ls-components-plus/compare/v1.8.0...v1.8.1) (2025-04-24)
+
+
+### Bug Fixes
+
+* **ls-component:** 🧩 优化docx头部展示 ([971f562](https://github.com/DTFN/ls-components-plus/commit/971f5629ba796899d7f79415179120b80a834411))
+
+# [1.8.0](https://github.com/DTFN/ls-components-plus/compare/v1.7.12...v1.8.0) (2025-04-24)
+
+
+### Bug Fixes
+
+* **ls-component:** 🧩 文档更新 ([3e7c468](https://github.com/DTFN/ls-components-plus/commit/3e7c468c19fe2c93766ef6f5a06a68606259a050))
+
+
+### Features
+
+* **ls-component:** 🚀 layout增加header，aside灰色样式 ([e4f67a2](https://github.com/DTFN/ls-components-plus/commit/e4f67a2ec9b58b44783acb6231647800d5085fc8))
+
+## [1.7.12](https://github.com/DTFN/ls-components-plus/compare/v1.7.11...v1.7.12) (2025-04-23)
+
+
+### Bug Fixes
+
+* **ls-component:** 🧩 拓展LSDescriptions column样式，增加label-width功能 ([470e805](https://github.com/DTFN/ls-components-plus/commit/470e805bb03ed0928c41cb9b06c5af035505b5a5))
+* **ls-component:** 🧩 文档更新 ([bec2a38](https://github.com/DTFN/ls-components-plus/commit/bec2a38d2f8f8cc1df039742a657870ad3929d29))
+
+## [1.7.11](https://github.com/DTFN/ls-components-plus/compare/v1.7.10...v1.7.11) (2025-04-17)
+
+
+### Bug Fixes
+
+* **ls-component:** 🧩 bar、line增加serieslabel自定义 ([e62ffd0](https://github.com/DTFN/ls-components-plus/commit/e62ffd069dac62dd625d7449e44cb2feebd25442))
+* **ls-component:** 🧩 chart支持valueFormatter扩展 ([73c1f7c](https://github.com/DTFN/ls-components-plus/commit/73c1f7c3e787b08e9c6507c17138d72ac84bffa5))
+* **ls-component:** 🧩 update ([a3297bc](https://github.com/DTFN/ls-components-plus/commit/a3297bcfa10baeb22e8f7c903b8042361c87803d))
+* **ls-component:** 🧩 update ([09694fb](https://github.com/DTFN/ls-components-plus/commit/09694fbb8b5cd3703b95446e0899d982672f35df))
+* **ls-component:** 🧩 update ([2477ef3](https://github.com/DTFN/ls-components-plus/commit/2477ef3bdff8ab14f4aedd0aebc47255ad7b605c))
+* **ls-component:** 🧩 代码优化 ([299d0cf](https://github.com/DTFN/ls-components-plus/commit/299d0cf202c9c0e12013fd1cb66277cfe8b12936))
+* **ls-component:** 🧩 代码优化 ([9b126d4](https://github.com/DTFN/ls-components-plus/commit/9b126d45c8e7fb93c9f0f72f3ca66dcb0282656b))
+* **ls-component:** 🧩 完善valueFormatter ([74b32d5](https://github.com/DTFN/ls-components-plus/commit/74b32d549db6c5ccc68dd0c738c62b257d9f3d51))
+* **ls-component:** 🧩 文档更新 ([1c568fa](https://github.com/DTFN/ls-components-plus/commit/1c568fa75d607c81c633f6739d5f49f588f74f6a))
+* **ls-component:** 🧩 文档更新 ([7051f8a](https://github.com/DTFN/ls-components-plus/commit/7051f8a4a3653daccdf5334b1f055823cfc30887))
+* **ls-component:** 🧩 文档更新 ([0ef94e6](https://github.com/DTFN/ls-components-plus/commit/0ef94e68a682a9517d33731e146efc7f78154ab4))
+* **ls-component:** 🧩 更新配置文件 ([5d0b8a4](https://github.com/DTFN/ls-components-plus/commit/5d0b8a4c1eec37a71ff408ce872814e12f384e94))
+* **ls-doc:** 🧩 更新chart自定义轴内容使用说明 ([a72afa4](https://github.com/DTFN/ls-components-plus/commit/a72afa4d9590d1ba3f44bf12daba0b2aaffb3986))
+
+## [1.7.11](https://github.com/DTFN/ls-components-plus/compare/v1.7.10...v1.7.11) (2025-04-17)
+
+
+### Bug Fixes
+
+* **ls-component:** 🧩 bar、line增加serieslabel自定义 ([e62ffd0](https://github.com/DTFN/ls-components-plus/commit/e62ffd069dac62dd625d7449e44cb2feebd25442))
+* **ls-component:** 🧩 chart支持valueFormatter扩展 ([73c1f7c](https://github.com/DTFN/ls-components-plus/commit/73c1f7c3e787b08e9c6507c17138d72ac84bffa5))
+* **ls-component:** 🧩 update ([a3297bc](https://github.com/DTFN/ls-components-plus/commit/a3297bcfa10baeb22e8f7c903b8042361c87803d))
+* **ls-component:** 🧩 update ([09694fb](https://github.com/DTFN/ls-components-plus/commit/09694fbb8b5cd3703b95446e0899d982672f35df))
+* **ls-component:** 🧩 update ([2477ef3](https://github.com/DTFN/ls-components-plus/commit/2477ef3bdff8ab14f4aedd0aebc47255ad7b605c))
+* **ls-component:** 🧩 代码优化 ([299d0cf](https://github.com/DTFN/ls-components-plus/commit/299d0cf202c9c0e12013fd1cb66277cfe8b12936))
+* **ls-component:** 🧩 代码优化 ([9b126d4](https://github.com/DTFN/ls-components-plus/commit/9b126d45c8e7fb93c9f0f72f3ca66dcb0282656b))
+* **ls-component:** 🧩 完善valueFormatter ([74b32d5](https://github.com/DTFN/ls-components-plus/commit/74b32d549db6c5ccc68dd0c738c62b257d9f3d51))
+* **ls-component:** 🧩 文档更新 ([7051f8a](https://github.com/DTFN/ls-components-plus/commit/7051f8a4a3653daccdf5334b1f055823cfc30887))
+* **ls-component:** 🧩 文档更新 ([0ef94e6](https://github.com/DTFN/ls-components-plus/commit/0ef94e68a682a9517d33731e146efc7f78154ab4))
+* **ls-component:** 🧩 更新配置文件 ([5d0b8a4](https://github.com/DTFN/ls-components-plus/commit/5d0b8a4c1eec37a71ff408ce872814e12f384e94))
+* **ls-doc:** 🧩 更新chart自定义轴内容使用说明 ([a72afa4](https://github.com/DTFN/ls-components-plus/commit/a72afa4d9590d1ba3f44bf12daba0b2aaffb3986))
+
+## [1.7.10](https://github.com/DTFN/ls-components-plus/compare/v1.7.9...v1.7.10) (2025-03-31)
+
+
+### Bug Fixes
+
+* **ls-component:** 🧩 pdf中文字体展示修复 ([28e610c](https://github.com/DTFN/ls-components-plus/commit/28e610cca1dd72cc93be29c21eb3b5ab33aa08d7))
+
+## [1.7.9](https://github.com/DTFN/ls-components-plus/compare/v1.7.8...v1.7.9) (2025-03-31)
+
+
+### Bug Fixes
+
+* **ls-component:** 🧩 修复lsmenu样式 ([6b58ac9](https://github.com/DTFN/ls-components-plus/commit/6b58ac9c3659bdc847644fde46fb8f4a331aae83))
+* **ls-component:** 🧩 修复pdf不识别中文汉字 ([93f976a](https://github.com/DTFN/ls-components-plus/commit/93f976a5623533cafda0f3c9c1f95ca547926789))
+* **ls-component:** 表单单个输入框时阻止回车刷新页面 ([3644b67](https://github.com/DTFN/ls-components-plus/commit/3644b6773aad010ba76300114330fbaa8017fed5))
+
+## [1.7.8](https://github.com/DTFN/ls-components-plus/compare/v1.7.7...v1.7.8) (2025-03-20)
+
+
+### Bug Fixes
+
+* **ls-component:** 🧩 icon对齐完善，xlsx多余高度处理 ([e934933](https://github.com/DTFN/ls-components-plus/commit/e9349333bf42fabc01c72b46cb9aba822e0dc725))
+
+## [1.7.7](https://github.com/DTFN/ls-components-plus/compare/v1.7.6...v1.7.7) (2025-03-20)
+
+
+### Bug Fixes
+
+* **ls-component:** 🧩 优化menu配置的icon对齐 ([9397477](https://github.com/DTFN/ls-components-plus/commit/9397477b7e8093ce6f53a90641257c8a6ee7204f))
+* **ls-doc:** 🧩 list文档更新 ([f91f8eb](https://github.com/DTFN/ls-components-plus/commit/f91f8ebe361083b875980e8019f6a468768006f6))
+* **ls-doc:** 🧩 文档更新 ([7fa27a0](https://github.com/DTFN/ls-components-plus/commit/7fa27a0b638fed41282051ffc34c21379edb67ef))
+
 ## [1.7.6](https://github.com/DTFN/ls-components-plus/compare/v1.7.5...v1.7.6) (2025-02-20)